edited content for cam3 page, renamed files to match all other pages, added new styles for it
