+++ client/src/pages/SignInFace.css
@@ -1,37 +1,19 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto&display=swap');
-body {
-  margin: 0;
-  font-family: 'Roboto', 'sans-serif';
-}
-.notification-container {
-    font-size: 14px;
-    box-sizing: border-box;
-    position: fixed;
-}
-.top-right {
-    top: 12px;
-    right: 12px;
-    transition: transform .6s ease-in-out;
-    animation: toast-in-right .7s;
+#main-container {
+    align-items: center;
+    justify-content: center;
 }
 
-.bottom-right {
-    bottom: 12px;
-    right: 12px;
-    transition: transform .6s ease-in-out;
-    animation: toast-in-right .7s;
+#main-jumboron {
+    align-items: center;
+    justify-content: center;
 }
 
-.top-left {
-    top: 12px;
-    left: 12px;
-    transition: transform .6s ease-in;
-    animation: toast-in-left .7s;
+#canvas {
+    border: solid 3px black;
+    width: 100%;
 }
 
-.bottom-left {
-    bottom: 12px;
-    left: 12px;
-    transition: transform .6s ease-in;
-    animation: toast-in-left .7s;
+#video {
+    border: solid 3px black;
+    width: 100%;
 }
